api/requirements.txt
asgiref==3.2.7
attrs==19.3.0
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
Django==3.0.5
djangorestframework==3.11.0
django-rest-multiple-models==2.1.3
django-pandas==0.6.1
Fiona==1.8.13
GDAL==3.0.4
geopandas==0.7.0
munch==2.5.0
numpy==1.17.0
pandas==1.0.3
psycopg2==2.8.5
pyproj==2.6.0
python-dateutil==2.8.1
pytz==2019.3
Shapely==1.7.0
six==1.12.0
sqlparse==0.3.1
google-api-python-client==1.8.2
oauth2client==4.1.2
jsonpickle==1.4.1
